refactor: test 패키지 domain에 맞춰 변경
